split Header into several files with the same namespace
