+++ srcs/app/config/analyser/ParserStatefulSet.cpp
@@ -1,4 +1,4 @@
-#include <parser/Parser.hpp>
+#include <parser/ParserStatefulSet.hpp>
 
 Parser::StatefulSet::StatefulSet(size_t line,
 				   t_parsing_state st,
