+++ srcs/app/config/analyser/ParserStatelessSet.cpp
@@ -1,4 +1,4 @@
-#include <parser/Parser.hpp>
+#include <parser/ParserStatelessSet.hpp>
 
 Parser::StatelessSet::StatelessSet(Engine *parser, API *config) :
 	config_(config),
