+++ srcs/app/config/analyser/ParsingState.cpp
@@ -1,4 +1,4 @@
-#include <parser/Parser.hpp>
+#include <parser/ParsingState.hpp>
 
 std::map<const std::string, t_parsing_state>
 								Parser::State::KeywordMapFactory_(void) {
