+++ srcs/app/test/config_analyser/ParserTest.cpp
@@ -5,7 +5,6 @@
 #include <parser/Lexer.hpp>
 #include <parser/Token.hpp>
 #include <parser/Preprocessor.hpp>
-#include <parser/Parser.hpp>
 #include <parser/ParserManager.hpp>
 #include <Config.hpp>
